% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_locking.R
\name{coherenceTest}
\alias{coherenceTest}
\title{Coherence for time series matrix}
\usage{
coherenceTest(x)
}
\arguments{
\item{x}{- a matrix, rows are genes, columns are times}
}
\value{
maximun coherence matrix between each pair of time series
}
\description{
Coherence for time series matrix
}
\examples{
#example 1
x = replicate(100, rnorm(5))
coherenceTest(x)

#example 2
x = rbind(sin(1:100),cos(1:100))
rownames(x)=c("sin","cos")
coherenceTest(x)

}
\references{

}

