% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_locking.R
\name{grangerTest}
\alias{grangerTest}
\title{Granger cause calculation for two vectors}
\usage{
grangerTest(x, p = 1)
}
\arguments{
\item{x}{- a matrix, rows are genes, columns are times}

\item{p}{- lag}
}
\value{
significant p-value with hypothesis that x granger cause y with lag=order
}
\description{
Granger cause calculation for two vectors
}
\examples{
#example 1
x = replicate(100, rnorm(5))
rownames(x)=paste("gene",1:nrow(x))
grangerTest(x, p=10)

#example 2
x = rbind(sin(1:100),cos(1:100))
rownames(x)=c("sin","cos")
grangerTest(x, p=4)

}
\references{
[1] http://www.r-bloggers.com/chicken-or-the-egg-granger-causality-for-the-masses/
[2] Thurman W.N. & Fisher M.E. (1988), Chickens, Eggs, and Causality, or Which Came First?, American Journal of Agricultural Economics, 237-238.
}

