% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{ModuleAnnotation}
\alias{ModuleAnnotation}
\title{Output the modules with GO annotation}
\usage{
ModuleAnnotation(colors)
}
\arguments{
\item{colors}{- numeric color vector return by wgcnaAnalysis.}
}
\value{
Table with the module color, gene names, GO annotation and statistics etc.
}
\description{
To export modules with GO.
}
\examples{

}

