% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{ModulePlots}
\alias{ModulePlots}
\title{Plot modules}
\usage{
ModulePlots(x, expr, mColors, thisColor = "blue", cut = 0.3)
}
\arguments{
\item{x}{- co-sychronization matrix returned by phaseLockingMatrix, grangerTest or coherenceTest}

\item{expr}{- expression profile}

\item{thisColor}{- the color of the module to be plotted}

\item{cut}{- the threshold for the co-sychronization matrix, extacting submatrix that is bigger than this cutoff, and plot the graph.}

\item{modulColors}{- color vector return by wgcnaAnalysis}
}
\value{
A graph and a heatmap
}
\description{
To visualize modules in graph and in heatmap
}
\examples{

}

