% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_locking.R
\name{phaseLockingMatrix}
\alias{phaseLockingMatrix}
\title{pahseLocking calculation for matrix}
\usage{
phaseLockingMatrix(x, n = 1, m = 1, method = "strobo")
}
\arguments{
\item{x}{- a matrix, rows are genes, columns are sorted by time}

\item{n}{- time period tatio}

\item{m}{- time period tatio}

\item{method}{- one in c(entrophy, gamma, strobo)}
}
\value{
list with 4 matrices

1. entropy rho 2. gamma  3. Strobo index lambda 4. Strobo index angle
}
\description{
pahseLocking is written for testing purposes.
if we are calculating phase locking for a large matrix,
running it per pair is slow
}
\examples{
x <- replicate(100, rnorm(20))
n=1
m=1
phaseLockingMatrix(x, n, m)
}
\references{
[1]
}

