% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{topGOanalysis}
\alias{topGOanalysis}
\title{Output the 2 modules overlaps.}
\usage{
topGOanalysis(colors, selectColor = "blue", identifier = "symbol",
  species = "Human", pvalue = 0.05, fileTree = TRUE, numSigTerm = 5)
}
\arguments{
\item{colors}{- numeric color vector return by wgcnaAnalysis.}

\item{selectColor}{- celcet the color of the module to perform Go enrichment analysis}

\item{identifier}{- default is "symbol". One in  c("Entrez", GenBank, Alias, Ensembl, GeneSymbol", "GeneName" and "UniGene"). Refer to topGO.}

\item{species}{- "Human" or "Mouse"}

\item{pvalue}{- export the GO terms with p-value smaller than this number}

\item{fileTree}{- TRUE if you wish to export the GO tree}

\item{numSigTerm}{- a parameter past to topGO function printGraph, determines how many top GO terms will be plotted in rectangle.}
}
\value{
GO enrichment result, and the tree structure GO plot with significant GO terms highlighted in red and orange colors. The pdf file of the tree will be saved in the working directory.
}
\description{
To export the number of genes shared between modules.
}
\examples{

}

