% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{wgcnaAnalysis}
\alias{wgcnaAnalysis}
\title{WGCNA modules}
\usage{
wgcnaAnalysis(x, softPower = 1)
}
\arguments{
\item{x}{- adjacency matrix returned by phaseLockingMatrix, grangerTest or coherenceTest}

\item{softPower}{- based on the plot of function wgcna_analysis_power, the 1st number that hists between 0.8-0.9 region should be picked as the soft-threshold. If none, use 1.}
}
\value{
3 plots.
}
\description{
As suggested by WGCNA, the 1st number that hists
between 0.8-0.9 region should be picked as the soft-threshold
}
\examples{
#example 1
x = replicate(1000, rnorm(1000))
wgcna_result=wgcna_analysis(x,5)

}
\references{
[1] Langfelder, P. & Horvath, S. WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 9, 559 (2008)
}

