% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{wgcnaOverlap}
\alias{wgcnaOverlap}
\title{Output the 2 modules overlaps.}
\usage{
wgcnaOverlap(modules1, modules2)
}
\arguments{
\item{modules1}{- modules return by wgcnaAnalysis.}

\item{modules2}{- modules return by wgcnaAnalysis.}
}
\value{
A figure and table with the number of overlapping genes, 
 correspondence (correspondence= number of overlap/ avergae module size),
 size of modules.
}
\description{
To export the number of genes shared between modules.
}
\examples{

}

