% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrap.R
\name{wgcnaPower}
\alias{wgcnaPower}
\title{soft-thresholding powers picking}
\usage{
wgcnaPower(x)
}
\arguments{
\item{x}{- adjacency matrix returned by phaseLockingMatrix, grangerTest or coherenceTest}
}
\value{
pdf file with R^2 plot and average connectivity
}
\description{
As suggested by WGCNA, the 1st number that hists
between 0.8-0.9 region should be picked as the soft-threshold
}
\examples{
#example 1
#be aware that this example data is generating a random matirx, so the plot does not match the real situation.  
x = replicate(1000, qunif(sort(runif(1000,min=0,max=1)), min = 0, max = 1))
x[lower.tri(x)] = t(x)[lower.tri(x)]
wgcnaPower(x)

}
\references{
[1] Langfelder, P. & Horvath, S. WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 9, 559 (2008)
}

